Calculator 수정 1. divide 메서드 double반환 타입으로 수정

+++ src/week3/day2_1003/Calculator.java
@@ -17,6 +17,6 @@ public class Calculator {
 
     public void divide(int a, int b) {
 
-        System.out.println(a / b);
+        System.out.println((double)a / b);
     }
 }
